Resolve circular dependency in AccessorsForOqlFilter
